Day 3 solution in Zig.
This was quite a learnig curve and I'm still not very happy with
the code (though I think this is more of a "choice-of-algorithm"
problem than my usage of Zig).

After storing the lines of input in an ArrayList I then do two
passes over the data:

Pass #1
Identify all of the numbers and whether they're adjacent to a
symbol (part 1 solution). As I go I compile a separate ArrayList
containing the numbers at each location in the grid.

Pass #2
Find the '*' symbols specifically and locate the adjacent numbers.
The actual gear ratio calculation became difficult because some
gears have the same number repeated. To address this I also had to
ID each number.

It wasn't a big problem in this particular case but in general it
seems like Zig's lack of a HashSet type could be a limitation going
forward.