Day 7 solution in Zig.
This turned out to be a right faff because I misread the comparison
logic for P2. Overall it was a rather fiddly one to solve, but it
felt nice to use the factory function idiom in Zig.

Another cool Zig discovery: you can get the strings corresponding
to enum values!