Day 10 solution in Zig.
This one was a big step up in difficulty over previous problems
so far this year. I've put in a bit of effort to refactor my
`Map` struct so that it's easier to read, but it's still a bit
of a behemoth.

For Part 2 I figured out pretty quickly that the approach must
be to consider the parity of number of cross-overs as we go
towards the edge of the map, but actually implementing this was
still rather awkward.