Day 15 solution in Zig.
Part 1 was straightforward, but Part 2 was a little bit of work.
The built-in HashMap type in Zig doesn't seem to support multiple
values per hash key, so it wasn't suitable.

My custom solution is a bit messy but gets the job done.

Interestingly reading the whole file in one hit was much faster
than a line-by-line approach (even though there's only one line).