Day 17 solution in C++.
Finally time to crack out the BFS skillz! The twist is that we have
to take extra care not to prune a path just because it reaches a
location and costs more: it could be that the steps taken leading
up to the location mean the future of the path is more favourable.

This is dealt with by tracking lowest cost to each location for each
possible combination of lead-up moves (I used a small hash function).

Given the work that went into Part 1 solving Part 2 was fairly
straightforward this time around.