Day 18 solution in C++.
I solved Part 1 quickly using a straightforward BFS but there was no
way this would generalise to Part 2!

I ended up doing some research and came across the Trapezoid formula.
This got me most of the way there, but it didn't account for the
thickness of the boundaries. I've explained this more in the code
comments but I was able to deduce how far off I would be and add the
boundaries back in.