Day 19 solution in C++.
It took me a bit of time to parse the input into a reasonable format
but once I had it Part 1 was straightforward.

Part 2 required the same kind of interval/bulk-counting method I
brute-forced my way past on an earlier day. Implementing this was
easy enough though the interval splitting logic was a bit fiddly.