Day 20 solution in C++.
A lot of Part 1 effort was just parsing the input into a useful
format. Once I had this it was a fairly simple queue-based
algorithm.

Part 2 was much more tricky. I ended up deep-diving into my input
to work backwards from the rx node and found a sequence of
conjunctions leading up to it. At this point I figured it must be
another periodic situation and just needed to work out the periods
with which the inputs nodes turned on.

It would be nice to have a more general solution to part 2 (I dont
know if this is practical).