Day 21 solution in C++.
Part 1 was a straightforward BFS situation.

I got pretty stuck on Part 2 and ended up looking for some hints.
In the end I went down the route of using my BFS algorithm to
generate data and then fitted a quadratic curve to that data.
There ended up being some error bars, but after a few attempts
I was able to interpolate the correct solution.

I'd like to revisit later and see if I can improve the efficiency
of my BFS - it does seem a bit unnecessary to store all of the
reachable points in a set.