Day 23 solution in C++.
A lot of code required today, but I had a strong suspicion from
the start that I would need to go down the graph route rather than
working directly with the grid, and this proved to be correct.

Part 1 was essentially implementing a directed graph. The hardest
part here was getting the edge weight calculation correct (I've
got a weird queue-based algorithm with recursion).

Part 2 needed the graph to be transformed into an undirected one
instead which, while fiddly, meant I could reuse most of my code.

It's particularly satisfying that given the directed -> undirected
change the same search algorithm works for both parts.