Day 24 solution in C++/Python.
Part 1 was a case of solving a pair of simultaneous equations
for each pair of hail stones.

For Part 2 I reduced the problem down to solving for 3 hail
stones, which lead to a (non-linear) system of 9 equations in
9 unknowns.

I'm not sure what algorithm to use to solve these, and don't have
a convenient C++ solving library to hand, so I ended up plugging
the hand-crafted equations into Python's SymPy module to solve.

Might come back and try a pure C++ solution later.